requirements.txt
python==3.11.5
requests==2.31.0
pillow==10.4.0
torch==2.4.1
torchvision==0.19.1
scikit-learn==1.5.1
numpy==1.26.4
matplotlib==3.9.2
tqdm==4.66.2
pandas==2.2.2